Rename filename with hypen instead of underscore
