+++ lib/core/providers-list.js
